17.1 CollectableItems. upload_file functionality created

+++ app_run/models.py
@@ -38,7 +38,7 @@ class Position(models.Model):
 class CollectibleItem(models.Model):
     name = models.CharField(max_length=255)
     uid = models.CharField(max_length=10)
-    latitude = models.DecimalField(max_digits=6, decimal_places=4)
-    longitude = models.DecimalField(max_digits=7, decimal_places=4)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     picture = models.URLField()
     value = models.PositiveSmallIntegerField()
